move apps to install.py, test folder namespaces
