+++ fabfile/test_tasks/test_task.py
@@ -1,0 +1,5 @@
+from fabric.api import *
+from fedora import install
+
+def test_install_nano():
+    install('nano')
